refactor: fix compilation errors, remove panics, and optimize clones
  Fixed all compilation errors in benchmarks/examples (unused imports, wrong
  module paths, outdated div syntax). Replaced panic!() calls in production
  code with proper Result returns - added AllocationError variant for this.

  Removed the unsafe block in page_reader.rs that had UB risk when decoder
  returned errors. The safe version with Vec::push compiles to the same thing
  anyway. Improved the unsafe code in aligned_buffer.rs to use the checked
  Layout::from_size_align instead of unchecked version.

  Went through and eliminated unnecessary clones, especially in hot paths:
  - Writer now takes &str instead of cloning device_id and measurement names
  - Changed chunk_reader to use Arc<str> for measurement_name (just bumps
    refcount instead of allocating)
  - Arrow conversion uses explicit Arc::clone for clarity
  - About ~20 clones removed total, documented the ones that need to stay

  Changed some function signatures to accept `impl Into<String>` so callers
  can pass either String or &str without explicit clones.

  Translated remaining Spanish comments/docs to English.

  All tests still passing (214 passing).

+++ src/encoding/chimp128.rs
@@ -108,7 +108,7 @@ impl Chimp128Encoder {
                 let length = bit_width - self.prev_leading - self.prev_trailing;
                 for i in (0..length).rev() {
                     self.buffer
-                        .push((xor >> (start + i as u8)) & 1 == 1);
+                        .push((xor >> (start + i)) & 1 == 1);
                 }
             } else if (leading >= self.prev_leading.saturating_sub(1)
                 && leading <= self.prev_leading + 1)
